sales-dashboard/requirements.txt
streamlit
pandas
matplotlib
pyyaml
pyarrow
fastparquet
numpy